+++ bridge_ui/src/hooks/useHandleCreateWrapped.tsx
@@ -36,6 +36,7 @@ import { isEVMChain } from "../utils/ethereum";
 import parseError from "../utils/parseError";
 import { signSendAndConfirm } from "../utils/solana";
 import { Alert } from "@material-ui/lab";
+import { postWithFees } from "../utils/terra";
 
 async function evm(
   dispatch: any,
@@ -119,10 +120,11 @@ async function terra(
       wallet.terraAddress,
       signedVAA
     );
-    const result = await wallet.post({
-      msgs: [msg],
-      memo: "Wormhole - Create Wrapped",
-    });
+    const result = await postWithFees(
+      wallet,
+      [msg],
+      "Wormhole - Create Wrapped"
+    );
     dispatch(
       setCreateTx({ id: result.result.txhash, block: result.result.height })
     );