+++ sdk/js/src/token_bridge/createWrapped.ts
@@ -22,16 +22,11 @@ export async function createWrappedOnTerra(
   walletAddress: string,
   signedVAA: Uint8Array
 ) {
-  return new MsgExecuteContract(
-    walletAddress,
-    tokenBridgeAddress,
-    {
-      submit_vaa: {
-        data: fromUint8Array(signedVAA),
-      },
+  return new MsgExecuteContract(walletAddress, tokenBridgeAddress, {
+    submit_vaa: {
+      data: fromUint8Array(signedVAA),
     },
-    { uluna: 1000 }
-  );
+  });
 }
 
 export async function createWrappedOnSolana(